# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Ledjer API
  version: 1.0.0
  description: >
    Cloudflare Worker (Hono) API for Ledjer - cash-only double-entry
    bookkeeping MVP for Indonesian UMKM. All money values are integer IDR
    (minor units). Protected endpoints require the session cookie
    (`__Host-ledjer_session` in production, `ledjer_session` elsewhere) and
    state-changing requests must send a same-origin `Origin` header.
servers:
  - url: https://ledjer.id
    description: Production
  - url: http://localhost:8788
    description: "Local dev (via Vite proxy: http://localhost:5173)"
...
paths:
...
  /api/transactions:
    get:
      tags: [Transactions]
      security: [{ cookieAuth: [] }]
      summary: List transactions (filters + pagination)
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: transactionType, in: query, schema: { $ref: "#/components/schemas/TransactionType" } }
        - { name: status, in: query, schema: { $ref: "#/components/schemas/TransactionStatus" } }
        - { name: fromDate, in: query, schema: { type: string, format: date } }
        - { name: toDate, in: query, schema: { type: string, format: date } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
      responses:
        "200":
          description: Transaction page
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items: { $ref: "#/components/schemas/Transaction" }
                  total: { type: integer }
    post:
      tags: [Transactions]
      security: [{ cookieAuth: [] }]
      summary: Post a transaction (idempotent by idempotencyKey)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [transactionType, transactionDate, cashAccountId, description, idempotencyKey]
              properties:
                transactionType: { $ref: "#/components/schemas/TransactionType" }
                transactionDate: { type: string, format: date, description: "YYYY-MM-DD; future dates (Asia/Jakarta) are rejected" }
                cashAccountId: { type: string }
                counterAccountId: { type: string, description: "Required for all types except purchase (inventory account is auto-resolved)" }
                amountIdr: { type: integer, exclusiveMinimum: 0, description: "Required for non-inventory types; for goods sale the total is computed from items" }
                description: { type: string, minLength: 1, maxLength: 200 }
                idempotencyKey: { type: string, minLength: 8, maxLength: 160 }
                items:
                  type: array
                  description: "Items for inventory transactions: required for purchase; optional for cash_in (goods sale)"
                  items:
                    type: object
                    required: [productId, quantity]
                    properties:
                      productId: { type: string }
                      quantity: { type: number, exclusiveMinimum: 0, description: "In product units; up to 3 decimals (e.g. 0.5 kg)" }
                      unitCostIdr: { type: integer, minimum: 0, description: "Purchase price per unit (purchase)" }
                      unitPriceIdr: { type: integer, minimum: 0, description: "Selling price per unit (goods sale)" }
      responses:
        "200":
          description: Posted transaction
          headers:
            Idempotent-Replay:
              schema: { type: string, enum: ["true"] }
              description: Set when the request replayed an earlier one
          content:
            application/json:
              schema:
                type: object
                properties:
                  transaction_id: { type: string }
                  transaction_number: { type: string }
                  journal_entry_id: { type: string }
                  status: { type: string, enum: [posted] }
                  replayed: { type: boolean }
        "400":
          description: Validation failed (account class mismatch, future date, ...)
        "409":
          description: Idempotency key reused with a different payload
        "429":
          description: Rate limited
  /api/transactions/{transactionId}:
    get:
      tags: [Transactions]
      security: [{ cookieAuth: [] }]
      summary: Transaction detail
      parameters:
        - { name: transactionId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                type: object
                properties:
                  transaction: { $ref: "#/components/schemas/Transaction" }
        "404":
          description: Not found in this book
  /api/transactions/{transactionId}/void:
    post:
      tags: [Transactions]
      security: [{ cookieAuth: [] }]
      summary: Void a posted transaction (reports then exclude it)
      parameters:
        - { name: transactionId, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: [string, "null"], maxLength: 500 }
      responses:
        "200":
          description: Voided transaction
          content:
            application/json:
              schema:
                type: object
                properties:
                  transaction: { $ref: "#/components/schemas/Transaction" }
        "409":
          description: Only posted transactions can be voided
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __Host-ledjer_session
  schemas:
...
    TransactionType:
      type: string
      enum: [cash_in, cash_out, transfer, owner_deposit, owner_withdrawal, purchase]
    TransactionStatus:
      type: string
      enum: [posted, voided]
...
    Transaction:
      type: object
      properties:
        id: { type: string }
        transaction_number: { type: string, example: "TRX-20260610-CD34" }
        transaction_type: { $ref: "#/components/schemas/TransactionType" }
        transaction_date: { type: string, format: date }
        description: { type: string }
        status: { $ref: "#/components/schemas/TransactionStatus" }
        amount_idr: { type: integer }
        cash_account_id: { type: string }
        counter_account_id: { type: string }
        cash_bank_account: { type: [string, "null"] }
        counter_account: { type: [string, "null"] }
        direction: { type: string, enum: [in, out, neutral] }
        user_id: { type: string }
        created_at: { type: integer }
        voided_at: { type: [integer, "null"] }
        void_reason: { type: [string, "null"] }
        items:
          type: [array, "null"]
          description: "Product items for inventory transactions; null otherwise"
          items:
            type: object
            properties:
              product_id: { type: string }
              product_code: { type: string }
              product_name: { type: string }
              quantity: { type: number }
              quantity_milli: { type: integer }
              unit_cost_idr: { type: number }
              cost_total_idr: { type: integer }
...